+++ src/services/plot/package.json
@@ -1,0 +1,3 @@
+{
+    "main": "plot.jsx"
+}
