+++ src/services/plot/plot.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import * as Plot from "react-chartjs-2";
-import { getConfig } from "./configPlot";
 
 export function renderPlot(state) {
     const { cols, data } = state;
@@ -27,4 +26,22 @@ export function renderPlot(state) {
             )
         );
     }
+}
+
+export function getConfig(cols) {
+    let config = {};
+    let values;
+
+    config.datasets = [];
+    values = cols.filter(o => o.type === "s");
+    if (values) config.labels = values[0].data.slice(1);
+    values = cols.filter(o => o.type === "n");
+    values.forEach(sequence =>
+        config.datasets.push({
+            label: sequence.data[0],
+            data: sequence.data.slice(1)
+        })
+    );
+
+    return config;
 }
