+++ src/services/spreadsheet/package.json
@@ -1,0 +1,3 @@
+{
+    "main": "spreadsheet.jsx"
+}
